--- configs/hendrycks_math/orch_v2_demo.toml
@@ -1,61 +1,0 @@
-# orch v2 showcase — small run, evals dominate wall-clock under the legacy
-# orchestrator (it pauses train rollouts during each eval epoch). Under v2 the
-# train/eval rollouts share a single concurrency limiter and overlap on each
-# eval boundary, so the total run finishes meaningfully faster.
-#
-# Designed for a 2-GPU box (1 train, 1 inference). 30 train steps with
-# eval.interval=5 → 6 eval epochs of AIME2024 with pass@4 (30 examples *
-# group_size=4 = 120 long-reasoning rollouts per epoch) — the eval workload
-# is heavy enough that the difference is obvious on wall clock.
-#
-# Run head-to-head to see the speedup (same TOML, only the
-# --orchestrator.experimental.use-orch-v2 flag flips):
-#
-#   # Legacy orchestrator (eval pauses train)
-#   uv run rl @ configs/hendrycks_math/orch_v2_demo.toml \
-#     --output-dir outputs/legacy --wandb.name legacy
-#
-#   # Orch v2 (eval overlaps with train)
-#   uv run rl @ configs/hendrycks_math/orch_v2_demo.toml \
-#     --orchestrator.experimental.use-orch-v2 \
-#     --output-dir outputs/orch_v2 --wandb.name orch_v2
-#
-# Watch wandb's wall-clock vs `step` axis on the two runs — v2 reaches step 30
-# noticeably ahead. Per-step reward should track between the two runs.
-
-max_steps = 30
-seq_len = 8192
-
-[wandb]
-project = "orch-v2-demo"
-name = "orch-v2-demo"
-
-[model]
-name = "deepseek-ai/DeepSeek-R1-Distill-Qwen-1.5B"
-
-[orchestrator]
-batch_size = 128
-group_size = 8
-
-[[orchestrator.train.env]]
-id = "math-env"
-name = "hendrycks-math"
-args = { dataset_name = "mikasenghaas/Sanity-Test-R1D-1.5B", dataset_subset = "default" }
-
-[orchestrator.eval]
-interval = 5
-
-[[orchestrator.eval.env]]
-id = "primeintellect/aime2024"
-name = "aime2024"
-group_size = 4
-
-[trainer] # Default trainer config
-
-[inference.model]
-max_model_len = 8192
-
-# Model not in MODEL_RENDERER_MAP — opt into DefaultRenderer (apply_chat_template).
-[orchestrator.renderer]
-name = "default"
-reasoning_parser = "think"
