+++ src/prime_rl/entrypoints/orchestrator.py
@@ -19,9 +19,9 @@ from prime_rl.utils.process import set_proc_title
 def main():
     set_proc_title("Orchestrator")
     config = cli(OrchestratorConfig)
-    from prime_rl.orchestrator.orchestrator import orchestrate
+    from prime_rl.orchestrator.orchestrator import run_orchestrator
 
-    asyncio.run(orchestrate(config))
+    asyncio.run(run_orchestrator(config))
 
 
 if __name__ == "__main__":
