+++ src/prime_rl/entrypoints/rl.py
@@ -192,16 +192,8 @@ def rl_local(config: RLConfig):
                 "orchestrator starts, otherwise rollouts will hang."
             )
 
-        # Start orchestrator process. Picks the v2 binary when
-        # ``orchestrator.experimental.use_orch_v2`` is set so we can flip
-        # implementations from the rl entrypoint without changing any TOML.
-        orchestrator_bin = "orchestrator-v2" if config.orchestrator.experimental.use_orch_v2 else "orchestrator"
-        orchestrator_cmd = [
-            orchestrator_bin,
-            "@",
-            (config_dir / ORCHESTRATOR_TOML).as_posix(),
-        ]
-        logger.info(f"Starting orchestrator process ({orchestrator_bin})")
+        orchestrator_cmd = ["orchestrator", "@", (config_dir / ORCHESTRATOR_TOML).as_posix()]
+        logger.info("Starting orchestrator process")
         logger.debug(f"Orchestrator start command: {' '.join(orchestrator_cmd)}")
         with open(log_dir / "orchestrator.log", "w") as log_file:
             orchestrator_process = Popen(
