+++ src/prime_rl/orchestrator/log_loop.py
@@ -11,9 +11,9 @@ from __future__ import annotations
 import asyncio
 import time
 
+from prime_rl.orchestrator.dispatcher import RolloutDispatcher
 from prime_rl.orchestrator.event_loop_lag import EventLoopLagMonitor
-from prime_rl.orchestrator_v2.dispatcher import RolloutDispatcher
-from prime_rl.orchestrator_v2.types import Policy
+from prime_rl.orchestrator.types import Policy
 from prime_rl.utils.async_utils import safe_cancel
 from prime_rl.utils.logger import get_logger
