+++ src/prime_rl/orchestrator/orchestrator.py
@@ -1,873 +1,660 @@
+"""Async-pipelined RL orchestrator.
+
+``Orchestrator`` is the only class that holds the whole picture: it owns the
+shared state (policy, progress, ckpt, monitor) and drives the pipeline loop.
+
+Components are deliberately single-purpose and only know about the deps they
+need:
+
+- ``RolloutDispatcher`` schedules rollouts and emits ``Rollout``\\ s on its queue.
+- ``TrainSink`` ingests train rollouts (tokenize → advantages + pre-filter →
+  post-filter), exposes ``pop_batch`` returning a ``TrainBatch``.
+- ``EvalSink`` ingests eval rollouts, exposes ``add`` returning an ``EvalBatch``
+  on epoch completion + ``build_metrics`` for log time.
+- ``MetricsBuilder`` builds the per-step train W&B dict from the popped batch
+  and orchestrator-side timings (called inline by the orchestrator).
+- ``WeightWatcher`` advances ``Policy`` and notifies the dispatcher.
+- ``IntervalLogger`` writes dispatcher gauges + event-loop lag on a time axis.
+
+None of these hold a reference to the orchestrator. The orchestrator wires
+them in ``setup()`` and drives them from ``main_loop()``.
+"""
+
+from __future__ import annotations
+
 import asyncio
 import ctypes
-import gc
 import os
 import time
 
 import tomli_w
 
-import prime_rl._compat  # noqa: F401 — patch ring_flash_attn compat before transitive import
-from prime_rl.orchestrator.advantage import compute_advantages
-from prime_rl.orchestrator.event_loop_lag import EventLoopLagMonitor
+import prime_rl._compat  # noqa: F401 — patch ring_flash_attn compat before transitive imports
+from prime_rl.configs.orchestrator import OrchestratorConfig
+from prime_rl.orchestrator.ckpt import setup_ckpt_manager
+from prime_rl.orchestrator.dispatcher import RolloutDispatcher
+from prime_rl.orchestrator.envs import EvalEnvs, TrainEnvs
+from prime_rl.orchestrator.eval_sink import EvalSink
+from prime_rl.orchestrator.filters import setup_filters
 from prime_rl.orchestrator.inference_metrics import InferenceMetricsCollector
-from prime_rl.orchestrator.patches import monkey_patch_chat_completion_logprobs, monkey_patch_oai_iterable_types
-from prime_rl.orchestrator.trajectories import (
-    backfill_rollout_tokens,
-    interleave_rollout,
-    offload_images_to_disk,
+from prime_rl.orchestrator.log_loop import IntervalLogger
+from prime_rl.orchestrator.metrics import MetricsBuilder
+from prime_rl.orchestrator.patches import (
+    monkey_patch_chat_completion_logprobs,
+    monkey_patch_oai_iterable_types,
 )
-from prime_rl.transport import TrainingBatch, TrainingSample, setup_training_batch_sender
+from prime_rl.orchestrator.train_sink import TrainSink
+from prime_rl.orchestrator.trajectories import offload_images_to_disk
+from prime_rl.orchestrator.types import EvalBatch, Policy, Progress, Rollout
+from prime_rl.orchestrator.utils import compute_teacher_logprobs, get_weight_dir, set_default_executor
+from prime_rl.orchestrator.vf_utils import get_seq_len, intercept_vf_logging, save_rollouts
+from prime_rl.orchestrator.watcher import WeightWatcher
+from prime_rl.trainer.model import setup_tokenizer
+from prime_rl.transport import TrainingBatch, setup_training_batch_sender
+from prime_rl.utils.async_utils import safe_cancel
+from prime_rl.utils.client import init_nccl_broadcast, setup_inference_pool
+from prime_rl.utils.heartbeat import Heartbeat
+from prime_rl.utils.logger import get_logger, setup_logger
+from prime_rl.utils.monitor import setup_monitor
 from prime_rl.utils.pathing import get_log_dir, get_rollout_dir, get_step_path
 from prime_rl.utils.usage_reporter import UsageReporter
-
-# This monkey patch is necessary to avoid Pydantic validating fields using typing.Iterable (e.g. in multimodal or tool call messages) lazily which leads to tokenization errors, for more info see https://github.com/PrimeIntellect-ai/prime-rl/pull/1249
-monkey_patch_oai_iterable_types()
-
-
-# This monkey patch is necessary to avoid heavy CPU overhead from constructing the OAI ChatCompletion Pydantic model with logprobs, for more info see https://github.com/PrimeIntellect-ai/prime-rl/pull/1189
-monkey_patch_chat_completion_logprobs()
-
-# Import environment before any other imports
-
-import pandas as pd
-import verifiers as vf
-from renderers.base import create_renderer
-
-from prime_rl.configs.orchestrator import OrchestratorConfig
-from prime_rl.orchestrator.buffer import Buffer
-from prime_rl.orchestrator.ckpt import Progress, setup_ckpt_manager
-from prime_rl.orchestrator.envs import EvalEnv, EvalEnvs, TrainEnvs
-from prime_rl.orchestrator.filters import apply_filters, setup_filters
-from prime_rl.orchestrator.scheduler import Scheduler
-from prime_rl.orchestrator.utils import (
-    compute_teacher_logprobs,
-    get_weight_dir,
-    print_benchmark,
-    set_default_executor,
-)
-from prime_rl.orchestrator.vf_utils import (
-    get_seq_len,
-    intercept_vf_logging,
-    save_rollouts,
-)
-from prime_rl.trainer.model import setup_tokenizer
-from prime_rl.utils.client import (
-    init_nccl_broadcast,
-    setup_inference_pool,
-)
-from prime_rl.utils.config import cli
-from prime_rl.utils.heartbeat import Heartbeat
-from prime_rl.utils.logger import setup_logger
-from prime_rl.utils.monitor import setup_monitor
-from prime_rl.utils.process import set_proc_title
 from prime_rl.utils.utils import (
     clean_exit,
     get_env_ids_to_install,
     install_env,
     resolve_latest_ckpt_step,
-    to_col_format,
 )
 
-# Hard wall-clock budget for the orchestrator's post-training cleanup. If the
-# graceful shutdown sequence (scheduler / inference pool / env teardown) is
-# still running after this many seconds, we force-exit the process so the run
-# pod terminates instead of sitting wedged forever. The training checkpoint
-# and artifacts are persisted *before* this point, so a forced exit is safe.
+monkey_patch_oai_iterable_types()
+monkey_patch_chat_completion_logprobs()
+
+
+# Hard wall-clock budget for the orchestrator's post-training cleanup.
+# Persist artifacts before this point, then force-exit if graceful shutdown
+# wedges (env-server ZMQ recv, vLLM admin aclose, etc).
 SHUTDOWN_TIMEOUT_S = 300
 
-# Maximum number of times to attempt generating a training batch when all
-# rollouts are filtered out. After this many attempts, the orchestrator crashes
-# rather than silently skipping training steps.
-MAX_EMPTY_BATCH_ATTEMPTS = 3
 
+class Orchestrator:
+    """``await Orchestrator(config).start()`` to run.
 
-@clean_exit
-async def orchestrate(config: OrchestratorConfig):
-    # Initialize the logger
-    logger = setup_logger(
-        config.log.level,
-        json_logging=config.log.json_logging,
-    )
-    intercept_vf_logging(logger="verifiers.serve", level="WARN")  # show logs from env clients
+    ``stop()`` from outside (or the orchestrator self-stops when the train
+    sink has driven ``progress.step`` to ``max_steps``). Cleanup happens at
+    the tail of ``start()``.
+    """
 
-    logger.info(f"Starting orchestrator ({config.training_mode})")
+    def __init__(self, config: OrchestratorConfig) -> None:
+        self.config = config
+        self.logger = setup_logger(config.log.level, json_logging=config.log.json_logging)
+        intercept_vf_logging(logger="verifiers.serve", level="WARN")
+        self.logger.info(f"Starting orchestrator ({config.training_mode})")
 
-    set_default_executor()
-    event_loop_lag_monitor = EventLoopLagMonitor()
-    event_loop_lag_monitor_task = asyncio.create_task(event_loop_lag_monitor.run())
+        if config.bench:
+            self.logger.warning(f"Running in benchmark mode (max_steps={config.max_steps})")
 
-    # Print warning if running in benchmark mode
-    if config.bench:
-        logger.warning(f"Running in benchmark mode (max_steps={config.max_steps})")
+        # Cheap synchronous setup. Heavy async work happens in ``setup()``.
+        self.progress = Progress()
+        self.ckpt_manager = setup_ckpt_manager(config.output_dir, config.ckpt)
+        self.policy = Policy(version=0, model_name="")
+        self.stopped = asyncio.Event()
 
-    # Save configs to output directory
-    config_dir = config.output_dir / "control"
-    config_dir.mkdir(parents=True, exist_ok=True)
-    with open(config_dir / "orch.toml", "wb") as f:
-        tomli_w.dump(config.model_dump(exclude_none=True, mode="json"), f)
+        # Populated during ``setup()``. Typing as ``Any`` keeps the class
+        # readable; the orchestrator's setup phase fills these in deterministic
+        # order so we never see ``None`` past ``setup()``.
+        self.tokenizer = None
+        self.renderer = None
+        self.mm_token_type_ids_mapping: dict[int, int] | None = None
+        self.student_inference = None
+        self.teacher_inference = None
+        self.monitor = None
+        self.heart: Heartbeat | None = None
+        self.usage_reporter: UsageReporter | None = None
+        self.inference_metrics: InferenceMetricsCollector | None = None
+        self.train_envs: TrainEnvs | None = None
+        self.eval_envs: EvalEnvs | None = None
+        self.sender = None
+        self.lora_name: str | None = None
+        self.resume_step: int | None = None
 
-    # Install environments
-    env_ids_to_install = set()
-    env_ids_to_install.update(get_env_ids_to_install(config.train.env))
-    if config.eval is not None:
-        env_ids_to_install.update(get_env_ids_to_install(config.eval.env))
+        # Components — built at the end of ``setup()``.
+        self.dispatcher: RolloutDispatcher | None = None
+        self.train_sink: TrainSink | None = None
+        self.eval_sink: EvalSink | None = None
+        self.metrics: MetricsBuilder | None = None
+        self.watcher: WeightWatcher | None = None
+        self.log_loop: IntervalLogger | None = None
 
-    for env_id in env_ids_to_install:
-        install_env(env_id, prerelease=config.env_install_prerelease)
+        # Background tasks for components with their own main loop.
+        self.component_tasks: list[asyncio.Task] = []
 
-    logger.info(f"Initializing tokenizer ({config.tokenizer})")
-    tokenizer = setup_tokenizer(config.tokenizer)
+    # ── lifecycle ──────────────────────────────────────────────────────────
 
-    # Set up student inference pool (required for all training modes).
-    logger.info(
-        f"Initializing student inference pool (base_url={', '.join(config.student.client.base_url)}, "
-        f"model={config.student.model.name})"
-    )
-    renderer, student_inference = await setup_student_inference_pool(
-        config=config,
-        tokenizer=tokenizer,
-        logger=logger,
-    )
+    async def setup(self) -> None:
+        """Async setup: install envs, load models/pools, prepare env workers,
+        resume from checkpoint, and construct components."""
+        config = self.config
+        set_default_executor()
 
-    # Token-id → modality marker (1 = image patch, 2 = video patch) used
-    # to build ``mm_token_type_ids`` per sample. The renderer is the
-    # single source of truth — it already knows its own special-token
-    # IDs (``<|image_pad|>`` etc.) from the tokenizer it owns, so the
-    # orchestrator never needs to load a separate ``AutoProcessor``.
-    # Text-only renderers expose an empty map (or no attribute).
-    mm_token_type_ids_mapping: dict[int, int] | None = (
-        getattr(renderer, "mm_token_type_id_map", None) if renderer is not None else None
-    )
-    if mm_token_type_ids_mapping == {}:
-        mm_token_type_ids_mapping = None
+        # Persist the resolved config alongside the run.
+        config_dir = config.output_dir / "control"
+        config_dir.mkdir(parents=True, exist_ok=True)
+        with open(config_dir / "orch.toml", "wb") as f:
+            tomli_w.dump(config.model_dump(exclude_none=True, mode="json"), f)
 
-    # Set up teacher inference pool (configured for opd or sft). Always MITO for
-    # simplicity - this also keeps external OAI-compatible teachers (PI inference,
-    # OpenAI) working as drop-in endpoints.
-    teacher_inference = None
-    if config.teacher is not None:
-        logger.info(
-            f"Initializing teacher inference pool (base_url={', '.join(config.teacher.client.base_url)}, "
-            f"model={config.teacher.model.name})"
+        # Install envs (train + eval).
+        env_ids_to_install = set(get_env_ids_to_install(config.train.env))
+        if config.eval is not None:
+            env_ids_to_install.update(get_env_ids_to_install(config.eval.env))
+        for env_id in env_ids_to_install:
+            install_env(env_id, prerelease=config.env_install_prerelease)
+
+        self.logger.info(f"Initializing tokenizer ({config.tokenizer})")
+        self.tokenizer = setup_tokenizer(config.tokenizer)
+
+        # Student inference pool.
+        self.logger.info(
+            f"Initializing student inference pool (base_url={', '.join(config.student.client.base_url)}, "
+            f"model={config.student.model.name})"
         )
-        teacher_inference = await setup_inference_pool(
-            config.teacher.client,
-            model_name=config.teacher.model.name,
-            train_client_type="openai_chat_completions",
+        self.renderer, self.student_inference = await setup_student_inference_pool(
+            config=config, tokenizer=self.tokenizer
+        )
+        self.mm_token_type_ids_mapping = (
+            getattr(self.renderer, "mm_token_type_id_map", None) if self.renderer is not None else None
+        )
+        if self.mm_token_type_ids_mapping == {}:
+            self.mm_token_type_ids_mapping = None
+
+        if config.teacher is not None:
+            self.logger.info(
+                f"Initializing teacher inference pool (base_url={', '.join(config.teacher.client.base_url)}, "
+                f"model={config.teacher.model.name})"
+            )
+            self.teacher_inference = await setup_inference_pool(
+                config.teacher.client,
+                model_name=config.teacher.model.name,
+                train_client_type="openai_chat_completions",
+            )
+
+        self.logger.info(f"Initializing monitor (wandb={config.wandb}, prime_monitor={config.prime_monitor})")
+        self.monitor = setup_monitor(
+            wandb_config=config.wandb,
+            prime_config=config.prime_monitor,
+            output_dir=config.output_dir,
+            tokenizer=self.tokenizer,
+            run_config=config,
+            keep_full_history=config.bench,
         )
 
-    # Setup monitor (may register the run and set RUN_ID in the environment)
-    logger.info(f"Initializing monitor (wandb={config.wandb}, prime_monitor={config.prime_monitor})")
-    monitor = setup_monitor(
-        wandb_config=config.wandb,
-        prime_config=config.prime_monitor,
-        output_dir=config.output_dir,
-        tokenizer=tokenizer,
-        run_config=config,
-        keep_full_history=config.bench,
-    )
+        if config.heartbeat is not None:
+            self.heart = Heartbeat(config.heartbeat.url)
 
-    # Read run_id AFTER setup_monitor so that newly registered runs are captured
-    run_id = os.getenv("RUN_ID", "")
+        usage_base_url = os.environ.get("PI_USAGE_BASE_URL")
+        usage_api_key = os.environ.get("PI_USAGE_API_KEY")
+        if usage_base_url and usage_api_key:
+            self.usage_reporter = UsageReporter()
 
-    # Usage reporter requires BOTH the base URL and the API key. Activating
-    # with only one set used to crash every POST inside httpx (None header
-    # value), so we now gate construction on both being present and log a
-    # clear warning when half-configured.
-    usage_base_url = os.environ.get("PI_USAGE_BASE_URL")
-    usage_api_key = os.environ.get("PI_USAGE_API_KEY")
-    if usage_base_url and usage_api_key:
-        usage_reporter = UsageReporter()
-    else:
-        if usage_base_url and not usage_api_key:
-            logger.warning("PI_USAGE_BASE_URL is set but PI_USAGE_API_KEY is missing; usage reporting disabled.")
-        usage_reporter = None
+        # Filters — train-only by design (eval rollouts are never filtered).
+        pre_filters = setup_filters(config.pre_batch_filters, vocab_size=self.tokenizer.vocab_size)
+        post_filters = setup_filters(config.post_batch_filters, vocab_size=self.tokenizer.vocab_size)
 
-    # Setup heartbeat (only on rank 0, orchestrator is single process)
-    heart = None
-    if config.heartbeat is not None:
-        logger.info("Initializing heartbeat")
-        heart = Heartbeat(config.heartbeat.url)
-
-    # Build rollout filters
-    rollout_filters = setup_filters(config.filters, vocab_size=tokenizer.vocab_size)
-
-    # Load environments
-    logger.info("Loading training environments")
-    train_envs = TrainEnvs(config.train.env)
-    if config.training_mode == "sft":
-        # Teacher rollouts don't need inference-side logprobs (the trainer
-        # reconstructs teacher tokens), and some external reasoning-model
-        # endpoints (e.g. openai/gpt-5*) reject the parameter.
-        for env in train_envs:
-            env.sampling_args.pop("logprobs", None)
-    logger.info(f"Loaded {len(train_envs)} training environment(s) ({', '.join(train_envs.names)})")
-
-    await train_envs.start(
-        log_dir=get_log_dir(config.output_dir.parent) / "envs" / "train",
-        log_level=config.log.vf_level,
-        json_logging=config.log.json_logging,
-    )
-    logger.success("Train environment(s) ready")
-
-    eval_envs: EvalEnvs | None = None
-    if config.eval:
-        logger.info("Loading eval environment(s)")
-        eval_envs = EvalEnvs(config.eval.env)
-        logger.info(f"Loaded {len(eval_envs)} eval environment(s) ({', '.join(eval_envs.names)})")
-
-        await eval_envs.start(
-            log_dir=get_log_dir(config.output_dir.parent) / "envs" / "eval",
+        # Envs.
+        self.logger.info("Loading training environments")
+        self.train_envs = TrainEnvs(config.train.env)
+        if config.training_mode == "sft":
+            for env in self.train_envs:
+                env.sampling_args.pop("logprobs", None)
+        self.logger.info(f"Loaded {len(self.train_envs)} training environment(s) ({', '.join(self.train_envs.names)})")
+        await self.train_envs.start(
+            log_dir=get_log_dir(config.output_dir.parent) / "envs" / "train",
             log_level=config.log.vf_level,
             json_logging=config.log.json_logging,
         )
-        logger.success("Eval environment(s) ready")
+        self.logger.success("Train environment(s) ready")
 
-    # Setup buffer
-    logger.info(f"Setting up buffer ({config.buffer})")
-    buffer = Buffer(train_envs, config.buffer)
+        if config.eval is not None:
+            self.logger.info("Loading eval environment(s)")
+            self.eval_envs = EvalEnvs(config.eval.env)
+            self.logger.info(f"Loaded {len(self.eval_envs)} eval environment(s) ({', '.join(self.eval_envs.names)})")
+            await self.eval_envs.start(
+                log_dir=get_log_dir(config.output_dir.parent) / "envs" / "eval",
+                log_level=config.log.vf_level,
+                json_logging=config.log.json_logging,
+            )
+            self.logger.success("Eval environment(s) ready")
 
-    # Get checkpoint manager
-    logger.info(f"Initializing checkpoint manager ({config.ckpt})")
-    ckpt_manager = setup_ckpt_manager(config.output_dir, config.ckpt)
-
-    checkpoint_step = None
-    if config.ckpt and config.ckpt.resume_step is not None and ckpt_manager is not None:
-        if config.ckpt.resume_step == -1:
-            checkpoint_step = resolve_latest_ckpt_step(ckpt_manager.ckpt_dir)
-        else:
-            checkpoint_step = config.ckpt.resume_step
-
-    scheduler = Scheduler(
-        train_envs=train_envs,
-        buffer=buffer,
-        student_inference=student_inference,
-        teacher_inference=teacher_inference,
-        max_inflight_rollouts=config.max_inflight_rollouts,
-        max_off_policy_steps=config.max_off_policy_steps,
-        tasks_per_minute=config.tasks_per_minute,
-        lora_name=config.student.model.lora.name if config.student.model.lora else None,
-        config=config,
-    )
-
-    # Wait for pools to be ready
-    logger.info("Waiting for student inference pool to be ready")
-    await student_inference.wait_for_ready(config.student.model.name)
-    logger.success("Student inference pool ready")
-    if teacher_inference is not None:
-        assert config.teacher is not None
-        logger.info("Waiting for teacher inference pool to be ready")
-        await teacher_inference.wait_for_ready(config.teacher.model.name)
-        logger.success("Teacher inference pool ready")
-
-    # Start inference metrics collector (requires W&B)
-    inference_metrics_collector = None
-    if config.wandb is not None and config.collect_inference_metrics:
-        inference_metrics_collector = InferenceMetricsCollector(student_inference.admin_clients)
-        await inference_metrics_collector.start()
-
-    # Set up weight broadcast backend (targets student inference)
-    logger.info(f"Initializing weight broadcast ({config.weight_broadcast})")
-    if config.weight_broadcast.type == "nccl":
-        await init_nccl_broadcast(
-            student_inference.admin_clients,
-            config.weight_broadcast.host,
-            config.weight_broadcast.port,
-            config.weight_broadcast.timeout,
-            inference_world_size=config.weight_broadcast.inference_world_size,
-            quantize_in_weight_transfer=config.weight_broadcast.quantize_in_weight_transfer,
-        )
-
-    # Setup training batch sender for sending training examples to trainer
-    logger.info(f"Initializing training batch sender ({config.rollout_transport})")
-    training_batch_sender = setup_training_batch_sender(config.output_dir, config.rollout_transport)
-
-    # Reset weights to base model if starting from scratch
-    progress = Progress()
-
-    if checkpoint_step is not None and ckpt_manager is not None:
-        ckpt_manager.load(progress, buffer, step=checkpoint_step)
-        logger.info(f"Resuming training from checkpoint step {checkpoint_step}")
-        scheduler.ckpt_step = progress.step  # Always resume from the latest checkpoint
-
-        # In NCCL mode, skip existence check - weights are broadcasted, not stored on disk
-        check_exists = config.weight_broadcast.type != "nccl"
-        wait_timeout = config.ckpt.wait_for_weights_timeout if config.ckpt else None
-        weights_path = get_weight_dir(
-            config.output_dir, scheduler.ckpt_step, check_exists=check_exists, wait_timeout=wait_timeout
-        )
-        lora_name = config.student.model.lora.name if config.student.model.lora else None
-        await student_inference.update_weights(weights_path, lora_name=lora_name, step=scheduler.ckpt_step)
-        if lora_name is not None:
-            student_inference.update_model_name(lora_name)
-            if scheduler.rollout_inference is student_inference:
-                scheduler.model_name = lora_name
-    else:
-        logger.info("Training from scratch")
-
-    # Iterate over dataset in batches
-    logger.info(f"Starting orchestrator loop (max_steps={config.max_steps or 'infinite'})")
-    is_first_step = True
-
-    while True:
-        # Check if this run has been evicted by the trainer
-        evicted_path = config.output_dir / "control" / "evicted.txt"
-        if evicted_path.exists():
-            reason = evicted_path.read_text().strip()
-            raise RuntimeError(f"Run evicted by trainer: {reason}")
-
-        # Capture ckpt_step once for consistency (it's updated inside the scheduler)
-        ckpt_step = scheduler.ckpt_step
-        scheduler.ckpt_step = ckpt_step
-
-        # Save checkpoint (if we are at an interval step and not at the first or last step)
-        is_last_step = config.max_steps is not None and progress.step == config.max_steps - 1
-        save_ckpt_time = 0
-        if (
-            ckpt_manager is not None
-            and (config.ckpt and config.ckpt.interval)
-            and not (is_first_step or is_last_step)
-            and progress.step % config.ckpt.interval == 0
-        ):
-            logger.info(f"Saving checkpoint at step {progress.step}")
-            save_ckpt_start_time = time.perf_counter()
-            ckpt_manager.save(progress, buffer, step=progress.step)
-            save_ckpt_time = time.perf_counter() - save_ckpt_start_time
-
-        # Break if we have reached the maximum number of steps
-        if config.max_steps and progress.step >= config.max_steps:
-            break
-
-        logger.info(f"Starting orchestrator step {progress.step}")
-        step_start_time = time.perf_counter()
-
-        # Run evals BEFORE training (blocking). Weight updates are paused via
-        # scheduler.checkpoint_ready during eval to ensure consistent weights.
-        # Each eval env has its own interval, so we check each independently.
-        envs_to_eval: list[EvalEnv] = []
-        if config.eval:
-            assert eval_envs is not None
-            if is_first_step and checkpoint_step is not None and config.eval.skip_eval_on_resume:
-                logger.info(f"Skipping online eval on resume (step={progress.step})")
+        # Resume.
+        if config.ckpt is not None and config.ckpt.resume_step is not None and self.ckpt_manager is not None:
+            if config.ckpt.resume_step == -1:
+                self.resume_step = resolve_latest_ckpt_step(self.ckpt_manager.ckpt_dir)
             else:
-                for eval_env in eval_envs:
-                    if progress.step % eval_env.config.interval == 0 and (
-                        progress.step > 0 or config.eval.eval_base_model
-                    ):
-                        envs_to_eval.append(eval_env)
+                self.resume_step = config.ckpt.resume_step
 
-        if envs_to_eval:
-            env_names = ", ".join(e.name for e in envs_to_eval)
-            logger.info(f"Running evals at step={progress.step} for {env_names}")
+        # Initial policy state — student model name; on resume below we may
+        # advance ``policy.version`` and the LoRA-adapter model name.
+        self.policy.model_name = self.student_inference.model_name
 
-            # Pause weight updates and re-scheduling of training rollouts during eval
-            # to avoid evaluating across different checkpoints and avoid congestion
-            scheduler.checkpoint_ready.clear()
+        # Wait for inference pools to be reachable.
+        self.logger.info("Waiting for student inference pool to be ready")
+        await self.student_inference.wait_for_ready(config.student.model.name)
+        self.logger.success("Student inference pool ready")
+        if self.teacher_inference is not None:
+            assert config.teacher is not None
+            self.logger.info("Waiting for teacher inference pool to be ready")
+            await self.teacher_inference.wait_for_ready(config.teacher.model.name)
+            self.logger.success("Teacher inference pool ready")
 
-            # For heavy eval workloads, it might be necessary additionally cancel in-flight training rollouts
-            if config.eval.cancel_inflight_rollouts_on_eval:
-                logger.info("Cancelling in-flight training rollouts before starting evals to avoid congestion.")
-                await scheduler.cancel_inflight_rollouts()
+        if config.wandb is not None and config.collect_inference_metrics:
+            self.inference_metrics = InferenceMetricsCollector(self.student_inference.admin_clients)
+            await self.inference_metrics.start()
 
-            eval_results = await asyncio.gather(
-                *[
-                    eval_env.evaluate(
-                        model_name=student_inference.model_name,
-                        get_client=student_inference.get_eval_client,
-                        step=progress.step,
-                        cache_salt=str(ckpt_step),
-                    )
-                    for eval_env in envs_to_eval
-                ]
+        self.logger.info(f"Initializing weight broadcast ({config.weight_broadcast})")
+        if config.weight_broadcast.type == "nccl":
+            await init_nccl_broadcast(
+                self.student_inference.admin_clients,
+                config.weight_broadcast.host,
+                config.weight_broadcast.port,
+                config.weight_broadcast.timeout,
+                inference_world_size=config.weight_broadcast.inference_world_size,
+                quantize_in_weight_transfer=config.weight_broadcast.quantize_in_weight_transfer,
             )
 
-            # Save eval rollouts to disk (fire-and-forget background thread)
-            eval_rollouts = [o for outputs in eval_results for o in outputs]
-            if eval_rollouts:
-                step_path = get_step_path(get_rollout_dir(config.output_dir), progress.step)
-                await asyncio.to_thread(
-                    save_rollouts, eval_rollouts, step_path / "eval_rollouts.jsonl", exclude_keys={"trajectory"}
-                )
+        self.logger.info(f"Initializing training batch sender ({config.rollout_transport})")
+        self.sender = setup_training_batch_sender(config.output_dir, config.rollout_transport)
 
-            # Resume weight updates
-            scheduler.checkpoint_ready.set()
+        self.lora_name = config.student.model.lora.name if config.student.model.lora else None
 
-        # Schedule generating the training batch. Retry on empty-after-filter
-        # batches so the trainer never receives an empty batch.
-        generate_completions_time = 0.0
-        train_rollouts: list[vf.RolloutOutput] = []
-        num_rollouts = 0
-        num_unique_examples = 0
-        n_trainable = 0
-        for attempt in range(MAX_EMPTY_BATCH_ATTEMPTS):
-            train_rollouts = await scheduler.generate_batch(step=progress.step)
-            generate_completions_time += scheduler.last_batch_generation_time
-
-            # Compute advantages (in-place)
-            num_rollouts = len(train_rollouts)
-            num_unique_examples = len({(r["env_name"], r["example_id"]) for r in train_rollouts})
-            await asyncio.to_thread(compute_advantages, train_rollouts, config.advantage)
-
-            # Apply rollout filters — sets rollout["filters"] and rollout["is_filtered"]
-            await asyncio.to_thread(apply_filters, rollout_filters, train_rollouts)
-
-            n_trainable = sum(1 for r in train_rollouts if not r["is_filtered"])
-            if n_trainable > 0:
-                break
-
-            if attempt == MAX_EMPTY_BATCH_ATTEMPTS - 1:
-                logger.error(
-                    f"Attempt {attempt + 1}/{MAX_EMPTY_BATCH_ATTEMPTS} at step {progress.step} "
-                    f"filtered out all {num_rollouts} rollouts - crashing orchestrator"
-                )
-                reason = (
-                    f"All {num_rollouts} rollouts were filtered out on "
-                    f"{MAX_EMPTY_BATCH_ATTEMPTS} consecutive attempts at step {progress.step}"
-                )
-                evicted_path = config.output_dir / "control" / "evicted.txt"
-                evicted_path.parent.mkdir(parents=True, exist_ok=True)
-                evicted_path.write_text(reason)
-                raise RuntimeError(reason)
-
-            logger.warning(
-                f"Attempt {attempt + 1}/{MAX_EMPTY_BATCH_ATTEMPTS} at step {progress.step} "
-                f"filtered out all {num_rollouts} rollouts - retrying batch generation"
+        # Restore from checkpoint (if resuming).
+        if self.resume_step is not None and self.ckpt_manager is not None:
+            self.ckpt_manager.load(self.progress, step=self.resume_step)
+            self.logger.info(f"Resuming orchestrator from checkpoint step {self.resume_step}")
+            check_exists = config.weight_broadcast.type != "nccl"
+            wait_timeout = config.ckpt.wait_for_weights_timeout if config.ckpt else None
+            weights_path = get_weight_dir(
+                config.output_dir, self.progress.step, check_exists=check_exists, wait_timeout=wait_timeout
             )
+            await self.student_inference.update_weights(weights_path, lora_name=self.lora_name, step=self.progress.step)
+            if self.lora_name is not None:
+                self.student_inference.update_model_name(self.lora_name)
+                self.policy.model_name = self.lora_name
+            self.policy.version = self.progress.step
+        else:
+            self.logger.info("Training from scratch")
 
-        trainable_ratio = n_trainable / num_rollouts
-        if trainable_ratio <= 0.1:
-            logger.warning(
-                f"Only {n_trainable}/{num_rollouts} rollouts in the batch are trainable "
-                f"({trainable_ratio:.1%}) - this can mean the tasks are too easy or too hard for the "
-                "model, consider reviewing the task difficulty of your environment(s)"
-            )
-
-        # Save train rollouts to disk (fire-and-forget background thread)
-        step_path = get_step_path(get_rollout_dir(config.output_dir), progress.step)
-        await asyncio.to_thread(
-            save_rollouts, train_rollouts, step_path / "train_rollouts.jsonl", exclude_keys={"trajectory"}
+        # ── Construct components — each gets only the deps it needs. ─────
+        self.dispatcher = RolloutDispatcher(
+            config=config,
+            train_envs=self.train_envs,
+            eval_envs=self.eval_envs,
+            student_inference=self.student_inference,
+            teacher_inference=self.teacher_inference,
+            policy=self.policy,
+            resume_step=self.resume_step,
+        )
+        self.metrics = MetricsBuilder(config)
+        self.train_sink = TrainSink(
+            config,
+            tokenizer=self.tokenizer,
+            renderer=self.renderer,
+            train_envs=self.train_envs,
+            mm_token_type_ids_mapping=self.mm_token_type_ids_mapping,
+            batch_size=config.batch_size,
+            token_batch_size=config.token_batch_size,
+            advantage_config=config.advantage,
+            pre_filters=pre_filters,
+            post_filters=post_filters,
+        )
+        self.eval_sink = EvalSink(eval_envs=self.eval_envs)
+        self.watcher = WeightWatcher(
+            config,
+            policy=self.policy,
+            student_inference=self.student_inference,
+            observers=[self.dispatcher],
+            lora_name=self.lora_name,
+            ckpt_step=self.progress.step,
+        )
+        self.log_loop = IntervalLogger(
+            dispatcher=self.dispatcher,
+            policy=self.policy,
+            interval=config.experimental.log_loop_interval,
         )
 
-        # Offload base64 images to disk to free memory. No-op for text-only
-        # rollouts (no ``data:image`` URLs to find); cheap to call always.
+    async def start(self) -> None:
+        """Run the orchestrator until shutdown. Drives setup, spawns the
+        background tasks, runs the main loop in this task, then cleans up."""
+        await self.setup()
+        config = self.config
+        self.logger.info(f"Starting orchestrator loop (max_steps={config.max_steps or 'infinite'})")
+        start_time = time.perf_counter()
+
+        # Spawn background loops (dispatcher schedules, watcher polls, log_loop
+        # emits gauges). The pipeline ``main_loop`` runs inline in this task.
+        assert self.dispatcher and self.watcher and self.log_loop
+        self.component_tasks = [
+            asyncio.create_task(self.dispatcher.start(), name="dispatcher"),
+            asyncio.create_task(self.watcher.start(), name="watcher"),
+            asyncio.create_task(self.log_loop.start(), name="log_loop"),
+        ]
+
+        try:
+            await self.main_loop()
+        finally:
+            self.logger.success(f"Orchestrator step loop done in {time.perf_counter() - start_time:.1f}s")
+            if self.eval_envs is not None and config.eval is not None:
+                self.logger.info("Running final evals")
+                await asyncio.gather(
+                    *(
+                        eval_env.evaluate(
+                            model_name=self.student_inference.model_name,
+                            get_client=self.student_inference.get_eval_client,
+                            step=self.progress.step,
+                            cache_salt=str(self.progress.step),
+                        )
+                        for eval_env in self.eval_envs
+                    ),
+                    return_exceptions=True,
+                )
+            assert self.monitor is not None
+            self.monitor.save_final_summary()
+            if self.ckpt_manager is not None:
+                self.logger.info("Writing final checkpoint")
+                self.ckpt_manager.save(self.progress, step=self.progress.step)
+            await self.shutdown()
+            self.logger.success("Orchestrator finished.")
+            try:
+                ctypes.CDLL("libc.so.6").malloc_trim(0)
+            except Exception as e:
+                get_logger().debug(f"malloc_trim(0) failed: {e}")
+
+    async def stop(self) -> None:
+        """Signal a graceful shutdown. ``start()`` will observe this on its
+        next ``main_loop`` iteration and drive the rest of the teardown."""
+        self.stopped.set()
+
+    # ── pipeline ──────────────────────────────────────────────────────────
+
+    async def main_loop(self) -> None:
+        """The pipeline driver. Consumes ``Rollout``\\ s from the dispatcher
+        (the atomic unit) and routes train vs eval to their respective sinks.
+
+        Both sinks own their own batch-boundary detection by counting
+        arrivals up to ``group_size`` (and ``num_examples * group_size`` for
+        eval epochs); their ``add()`` return values signal "batch ready"
+        directly, so the orchestrator doesn't peek at any rollout flag.
+        """
+        assert self.dispatcher and self.train_sink and self.eval_sink
+        while not self.stopped.is_set():
+            try:
+                rollout: Rollout = await asyncio.wait_for(self.dispatcher.out_q.get(), timeout=0.5)
+            except asyncio.TimeoutError:
+                continue
+
+            if rollout.kind == "eval":
+                eval_batch = self.eval_sink.add(rollout)
+                if eval_batch is not None:
+                    self.log_eval_batch(eval_batch)
+                continue
+
+            batch_ready = await self.train_sink.add(rollout)
+            while batch_ready and not self.stopped.is_set():
+                await self.process_one_step()
+                # Another batch may have accumulated while we were processing
+                # this one; keep popping until the buffer is below threshold.
+                batch_ready = self.train_sink.is_batch_done()
+
+    async def process_one_step(self) -> None:
+        """Drive one shipping step. The train sink has done all per-rollout /
+        per-group / per-batch processing already; the orchestrator handles the
+        I/O concerns (ship to trainer, save rollouts, monitor log, heartbeat,
+        usage reporter, ckpt save, step++)."""
+        assert self.train_sink and self.metrics and self.dispatcher and self.monitor and self.sender
+        config = self.config
+        step = self.progress.step
+
+        save_ckpt_time = await self.maybe_save_ckpt(step)
+
+        if config.max_steps is not None and step >= config.max_steps:
+            self.logger.success(f"Reached max_steps={config.max_steps}, signaling shutdown")
+            self.stopped.set()
+            return
+
+        self.logger.info(f"Starting orchestrator step {step}")
+        step_start = time.perf_counter()
+
+        # Pop the next ready batch off the train sink. ``pop_batch`` returns
+        # the raw cohort + the trainer-bound samples + per-rollout counters;
+        # metrics are built at log time below with the final timings.
+        batch = self.train_sink.pop_batch()
+
+        if batch.result.n_trainable == 0:
+            self.logger.warning(
+                f"Step {step}: post-batch filters dropped all {len(batch.rollouts)} rollouts. Trying again."
+            )
+            return
+        if batch.result.n_trainable / len(batch.rollouts) <= 0.1:
+            self.logger.warning(
+                f"Only {batch.result.n_trainable}/{len(batch.rollouts)} rollouts in the batch are trainable "
+                f"({batch.result.n_trainable / len(batch.rollouts):.1%}) — consider reviewing task difficulty / filter config"
+            )
+
+        # Persist rollouts to disk (cheap, background thread).
+        step_path = get_step_path(get_rollout_dir(config.output_dir), step)
+        await asyncio.to_thread(
+            save_rollouts, batch.rollouts, step_path / "train_rollouts.jsonl", exclude_keys={"trajectory"}
+        )
+
+        # Offload base64 image bytes to disk for memory hygiene (no-op for text-only).
         offload_start = time.perf_counter()
-        num_offloaded = offload_images_to_disk(train_rollouts, config.output_dir)
+        num_offloaded = offload_images_to_disk(batch.rollouts, config.output_dir)
         if num_offloaded:
-            logger.info(
+            self.logger.info(
                 f"Offloaded {num_offloaded} unique images to disk in {time.perf_counter() - offload_start:.2f}s"
             )
 
-        # Convert rollouts to training samples
-        parallel_preprocess_start = time.perf_counter()
-
-        # We only expect to backfill tokens for training_mode=sft against an
-        # external teacher API (OpenAI/etc.), which returns no token IDs —
-        # reconstruct via tokenizer/renderer. The vLLM-served paths (RL/OPD
-        # renderer + MITO, and training_mode=sft against a local vLLM teacher)
-        # already populate tokens via prompt_token_ids/token_ids, so we
-        # short-circuit the 256-way fanout.
-        needs_backfill = any(step["tokens"] is None for rollout in train_rollouts for step in rollout["trajectory"])
-        if needs_backfill:
-            logger.info(
-                "Backfilling tokens for rollout trajectories (expected for training_mode=sft against an external teacher API)"
-            )
-            await asyncio.gather(
-                *(
-                    asyncio.to_thread(
-                        backfill_rollout_tokens,
-                        rollout,
-                        tokenizer,
-                        renderer=renderer,
-                    )
-                    for rollout in train_rollouts
-                )
-            )
-
-        # Process rollouts in parallel
-        results = await asyncio.gather(
-            *(
-                asyncio.to_thread(interleave_rollout, r, mm_token_type_ids_mapping=mm_token_type_ids_mapping)
-                for r in train_rollouts
-            )
-        )
-
-        # Collect results and assign advantages. Metrics are computed over all
-        # rollouts; only non-filtered samples are sent to the trainer.
-        train_examples: list[TrainingSample] = []
-        rollout_prefill_lens: list[int] = []
-        rollout_decode_lens: list[int] = []
-        rollout_samples_per_rollout: list[int] = []
-        num_prefill_tokens = 0
-        num_decode_tokens = 0
-        for rollout, samples in zip(train_rollouts, results):
-            rollout_prefill_tokens = 0
-            rollout_decode_tokens = 0
-            if samples is None:
-                samples = []
-            rollout_samples_per_rollout.append(len(samples))
-            for sample in samples:
-                sample.advantage = rollout["advantage"]
-                sample.reward = rollout["reward"]
-                sample.env_name = rollout["env_name"]
-                sample.training_mode = config.training_mode
-                sample_decode_tokens = sum(sample.completion_mask)
-                sample_prefill_tokens = len(sample.prompt_ids) + len(sample.completion_mask) - sample_decode_tokens
-                rollout_decode_tokens += sample_decode_tokens
-                rollout_prefill_tokens += sample_prefill_tokens
-                if not rollout["is_filtered"]:
-                    train_examples.append(sample)
-            rollout_prefill_lens.append(rollout_prefill_tokens)
-            rollout_decode_lens.append(rollout_decode_tokens)
-            num_prefill_tokens += rollout_prefill_tokens
-            num_decode_tokens += rollout_decode_tokens
-
-        parallel_preprocess_time = time.perf_counter() - parallel_preprocess_start
-        logger.debug(
-            f"Converted {len(train_rollouts)} rollouts ({num_unique_examples} unique examples) "
-            f"to {len(train_examples)} training examples"
-        )
-
-        # Compute teacher logprobs (opd only - sft trains on teacher tokens directly)
-        teacher_logprobs_time = 0
-        if config.training_mode == "opd" and teacher_inference is not None:
+        # Teacher logprobs (opd only).
+        teacher_logprobs_time = 0.0
+        if config.training_mode == "opd" and self.teacher_inference is not None:
             assert config.teacher is not None
-            logger.info(f"Computing teacher logprobs for {len(train_examples)} training examples")
-            teacher_logprobs_start_time = time.perf_counter()
+            t = time.perf_counter()
             teacher_logprobs_list = await compute_teacher_logprobs(
-                clients=teacher_inference.train_clients,
+                clients=self.teacher_inference.train_clients,
                 model_name=config.teacher.model.name,
-                samples=train_examples,
+                samples=batch.samples,
             )
-            for train_example, teacher_logprobs in zip(train_examples, teacher_logprobs_list):
-                train_example.teacher_logprobs = teacher_logprobs
-            teacher_logprobs_time = time.perf_counter() - teacher_logprobs_start_time
-            logger.debug(f"Computed teacher logprobs in {teacher_logprobs_time:.2f}s")
+            for ex, lp in zip(batch.samples, teacher_logprobs_list):
+                ex.teacher_logprobs = lp
+            teacher_logprobs_time = time.perf_counter() - t
 
-        training_batch = TrainingBatch(
-            examples=train_examples,
-            step=progress.step,
+        await self.wait_barrier(step)
+
+        # Ship to trainer.
+        await self.sender.send(TrainingBatch(examples=batch.samples, step=step))
+
+        # Build + ship metrics with the final post-barrier timings.
+        step_time = time.perf_counter() - step_start
+        metrics = self.metrics.build(
+            step=step,
+            rollouts=batch.rollouts,
+            result=batch.result,
+            progress=self.progress,
+            dispatcher_gauges=self.dispatcher.gauges(),
+            dispatcher_drain=self.dispatcher.drain_metrics(),
+            step_time=step_time,
+            save_ckpt_time=save_ckpt_time,
+            teacher_logprobs_time=teacher_logprobs_time,
+            pre_filter_seen=self.train_sink.pre_filter_seen,
+            pre_filter_dropped=self.train_sink.pre_filter_dropped,
+            pre_filter_dropped_by_name=dict(self.train_sink.pre_filter_dropped_by_name),
         )
-
-        await training_batch_sender.send(training_batch)
-
-        step_time = time.perf_counter() - step_start_time
-
-        # Gather metrics in dataframes
-        results_df = pd.DataFrame(
-            {
-                "example_id": [rollout["example_id"] for rollout in train_rollouts],
-                "env_name": [rollout["env_name"] for rollout in train_rollouts],
-                "reward": [rollout["reward"] for rollout in train_rollouts],
-                "is_truncated": [rollout["is_truncated"] for rollout in train_rollouts],
-                "is_filtered": [rollout["is_filtered"] for rollout in train_rollouts],
-                "stop_condition": [rollout.get("stop_condition") for rollout in train_rollouts],
-                "seq_len": [get_seq_len(rollout) for rollout in train_rollouts],
-                "prefill_len": rollout_prefill_lens,
-                "decode_len": rollout_decode_lens,
-                "samples_per_rollout": rollout_samples_per_rollout,
-                "num_turns": [len(rollout["trajectory"]) for rollout in train_rollouts],
-            }
-        )
-
-        # Separate DataFrames for env reward function metrics, filter flags, and per-rollout timings
-        # to avoid column name collisions
-        metrics_df = pd.DataFrame([rollout["metrics"] for rollout in train_rollouts])
-        filter_df = pd.DataFrame([rollout["filters"] for rollout in train_rollouts])
-        timing_df = pd.DataFrame(
-            [
-                {
-                    "total": rollout["timing"]["total"],
-                    "setup": rollout["timing"]["setup"]["duration"],
-                    "generation": rollout["timing"]["generation"]["duration"],
-                    "model": rollout["timing"]["model"]["duration"],
-                    "env": rollout["timing"]["env"]["duration"],
-                    "scoring": rollout["timing"]["scoring"]["duration"],
-                    "overhead": rollout["timing"]["overhead"],
-                }
-                for rollout in train_rollouts
-            ]
-        )
-
-        # Update progress metrics
-        num_tokens = int(results_df.seq_len.sum())
-        progress.total_tokens += num_tokens
-        progress.total_samples += num_rollouts
-        progress.total_problems += num_unique_examples
-
-        def compute_solve_rates(df):
-            """Compute solve_none, solve_all, effective_batch_size for a set of rollouts."""
-            reward_per_problem = df.groupby(["env_name", "example_id"]).reward.sum()
-            solve_none = (reward_per_problem == 0).mean()
-            solve_all = (reward_per_problem == config.group_size).mean()
-            return solve_none, solve_all, 1 - solve_none - solve_all
-
-        # Group by (env_name, example_id) to average across rollouts within each problem
-        by_example = results_df.groupby(["env_name", "example_id"])
-
-        solve_none, solve_all, effective_batch_size = compute_solve_rates(results_df)
-        to_log = {
-            # Progress metrics
-            "progress/tokens": num_tokens,
-            "progress/prefill_tokens": num_prefill_tokens,
-            "progress/decode_tokens": num_decode_tokens,
-            "progress/samples": num_rollouts,
-            "progress/problems": num_unique_examples,
-            "progress/total_tokens": progress.total_tokens,
-            "progress/total_samples": progress.total_samples,
-            "progress/total_problems": progress.total_problems,
-            # Sequence length metrics
-            "seq_len/all/mean": by_example.seq_len.mean().mean(),
-            "seq_len/all/max": by_example.seq_len.mean().max(),
-            "seq_len/all/min": by_example.seq_len.mean().min(),
-            "prefill_len/all/mean": by_example.prefill_len.mean().mean(),
-            "prefill_len/all/max": by_example.prefill_len.mean().max(),
-            "prefill_len/all/min": by_example.prefill_len.mean().min(),
-            "decode_len/all/mean": by_example.decode_len.mean().mean(),
-            "decode_len/all/max": by_example.decode_len.mean().max(),
-            "decode_len/all/min": by_example.decode_len.mean().min(),
-            "is_truncated/all/mean": by_example.is_truncated.mean().mean(),
-            "is_truncated/all/max": by_example.is_truncated.mean().max(),
-            "stop_condition/all/generation_truncated": (
-                results_df.is_truncated & (results_df.stop_condition != "prompt_too_long")
-            ).mean(),
-            **{
-                f"stop_condition/all/{sc}": rate
-                for sc, rate in results_df.stop_condition.dropna().value_counts(normalize=True).items()
-            },
-            "samples_per_rollout/all/mean": by_example.samples_per_rollout.mean().mean(),
-            "samples_per_rollout/all/max": by_example.samples_per_rollout.mean().max(),
-            "samples_per_rollout/all/min": by_example.samples_per_rollout.mean().min(),
-            "num_turns/all/mean": by_example.num_turns.mean().mean(),
-            "num_turns/all/max": by_example.num_turns.mean().max(),
-            "num_turns/all/min": by_example.num_turns.mean().min(),
-            **{
-                f"timing/all/{key}/{stat}": getattr(
-                    timing_df[key].groupby([results_df.env_name, results_df.example_id]).mean(),
-                    stat,
-                )()
-                for key in timing_df.columns
-                for stat in ("mean", "max", "min")
-            },
-            # Train reward
-            "reward/all/mean": by_example.reward.mean().mean(),
-            "reward/all/max": by_example.reward.mean().max(),
-            "reward/all/min": by_example.reward.mean().min(),
-            # Solve / batch metrics
-            "solve_none/all": solve_none,
-            "solve_all/all": solve_all,
-            "effective_batch_size/all": effective_batch_size,
-            **{f"batch/{env}": r for env, r in results_df.env_name.value_counts(normalize=True).items()},
-            # Time metrics
-            "time/step": step_time,
-            "time/generate_completions": generate_completions_time,
-            "time/teacher_logprobs": teacher_logprobs_time,
-            "time/save_ckpt": save_ckpt_time,
-            "time/parallel_preprocess": parallel_preprocess_time,
-            # Scheduler metrics
-            **scheduler.get_metrics(),
-            # Buffer metrics
-            **buffer.get_metrics(),
-            # Event loop lag metrics
-            **event_loop_lag_monitor.get_metrics(),
-            # Rollout filter metrics (detection rate per filter + overall drop rate)
-            "filters/all/is_filtered": results_df.is_filtered.astype(float).mean(),
-            **{f"filters/all/{name}": filter_df[name].astype(float).mean() for name in filter_df.columns},
-            # W&B axis
-            "step": progress.step,
-        }
-
-        # Per-env metrics
-        per_env_columns = [
-            "seq_len",
-            "prefill_len",
-            "decode_len",
-            "is_truncated",
-            "samples_per_rollout",
-            "num_turns",
-        ]
-
-        for env, env_df in results_df.groupby("env_name"):
-            env_by_example = env_df.groupby("example_id")
-            for col in per_env_columns:
-                to_log[f"{col}/{env}/mean"] = env_by_example[col].mean().mean()
-                to_log[f"{col}/{env}/max"] = env_by_example[col].mean().max()
-                if col != "is_truncated":
-                    to_log[f"{col}/{env}/min"] = env_by_example[col].mean().min()
-            env_timing_df = timing_df.loc[env_df.index]
-            for key in timing_df.columns:
-                per_example = env_timing_df.groupby(env_df["example_id"])[key].mean()
-                to_log[f"timing/{env}/{key}/mean"] = per_example.mean()
-                to_log[f"timing/{env}/{key}/max"] = per_example.max()
-                to_log[f"timing/{env}/{key}/min"] = per_example.min()
-            to_log[f"reward/{env}/mean"] = env_by_example.reward.mean().mean()
-            to_log[f"reward/{env}/max"] = env_by_example.reward.mean().max()
-            to_log[f"reward/{env}/min"] = env_by_example.reward.mean().min()
-            solve_none, solve_all, effective_batch_size = compute_solve_rates(env_df)
-            to_log[f"solve_none/{env}"] = solve_none
-            to_log[f"solve_all/{env}"] = solve_all
-            to_log[f"effective_batch_size/{env}"] = effective_batch_size
-            to_log[f"stop_condition/{env}/generation_truncated"] = (
-                env_df.is_truncated & (env_df.stop_condition != "prompt_too_long")
-            ).mean()
-            for sc, rate in env_df.stop_condition.dropna().value_counts(normalize=True).items():
-                to_log[f"stop_condition/{env}/{sc}"] = rate
-            env_metrics_df = metrics_df.loc[env_df.index]
-            for metric in metrics_df.columns:
-                to_log[f"metrics/{env}/{metric}"] = env_metrics_df.groupby(env_df["example_id"])[metric].mean().mean()
-            to_log[f"filters/{env}/is_filtered"] = env_df.is_filtered.astype(float).mean()
-            env_filter_df = filter_df.loc[env_df.index]
-            for name in filter_df.columns:
-                to_log[f"filters/{env}/{name}"] = env_filter_df[name].astype(float).mean()
-
-        # Log metrics to monitor(s)
-        monitor.log(to_log, step=progress.step)
-
-        # Log samples to monitor(s) if enabled.
-        monitor.log_samples(train_rollouts, step=progress.step)
-
-        # Log distributions (rewards, advantages) if enabled
-        monitor.log_distributions(
+        self.monitor.log(metrics, step=step)
+        self.monitor.log_samples(batch.rollouts, step=step)
+        self.monitor.log_distributions(
             distributions={
-                "rewards": [r["reward"] for r in train_rollouts],
-                "advantages": [r["advantage"] for r in train_rollouts],
+                "rewards": [r["reward"] for r in batch.rollouts],
+                "advantages": [r["advantage"] for r in batch.rollouts],
             },
-            step=progress.step,
+            step=step,
         )
 
-        if usage_reporter and run_id:
-            usage_reporter.report_training_usage(
-                run_id=run_id,
-                step=progress.step,
-                tokens=num_prefill_tokens + num_decode_tokens,
-            )
-
-        reward_mean = by_example.reward.mean().mean()
-        step_message = f"Step {progress.step} | Time: {step_time:.2f}s | Reward: {reward_mean:.4f} | Seq. Length: {by_example.seq_len.mean().mean():.1f} tokens/sample | Max. Off-Policy Level: {scheduler.max_off_policy_level}"
-        logger.success(step_message)
-
-        # Increment step
-        progress.step += 1
-        is_first_step = False
-
-        # Free large per-step objects to prevent memory accumulation
-        del train_rollouts, train_examples, training_batch
-        del results_df, metrics_df
-        gc.collect()
-        # Return free glibc heap pages to the OS. numpy/pandas allocate array data
-        # via malloc (outside Python's allocator), so gc.collect() alone doesn't
-        # reclaim the RSS. malloc_trim(0) forces glibc to return freed pages.
-        try:
-            ctypes.CDLL("libc.so.6").malloc_trim(0)
-        except Exception as e:
-            logger.warning(f"malloc_trim(0) failed - RSS may grow unboundedly: {e}")
-
-        event_loop_lag_monitor.reset()
-
-        # Send heartbeat if configured
-        if heart is not None:
-            heart.beat()
-
-    if config.eval and eval_envs is not None:
-        logger.info("Running final evals")
-        eval_results = await asyncio.gather(
-            *[
-                eval_env.evaluate(
-                    model_name=student_inference.model_name,
-                    get_client=student_inference.get_eval_client,
-                    step=progress.step,
-                    cache_salt=str(ckpt_step),
+        if self.usage_reporter is not None:
+            run_id = os.getenv("RUN_ID", "")
+            if run_id:
+                self.usage_reporter.report_training_usage(
+                    run_id=run_id,
+                    step=step,
+                    tokens=batch.result.num_prefill_tokens + batch.result.num_decode_tokens,
                 )
-                for eval_env in eval_envs
-            ]
+        if self.heart is not None:
+            self.heart.beat()
+
+        # Update progress totals.
+        num_rollouts = len(batch.rollouts)
+        num_unique_examples = len({(r["env_name"], r["example_id"]) for r in batch.rollouts})
+        num_tokens = sum(get_seq_len(r) for r in batch.rollouts)
+        self.progress.total_tokens += num_tokens
+        self.progress.total_samples += num_rollouts
+        self.progress.total_problems += num_unique_examples
+
+        reward_mean = float(sum(r["reward"] for r in batch.rollouts) / max(num_rollouts, 1))
+        self.logger.success(
+            f"Step {step} | Time: {step_time:.2f}s | Reward: {reward_mean:.4f} | "
+            f"Seq. Length: {num_tokens / max(num_rollouts, 1):.1f} tokens/sample | "
+            f"Trainable: {batch.result.n_trainable}/{num_rollouts} | "
+            f"Async Level: {step - self.policy.version} | "
+            f"Max. Off-Policy Level: {self.dispatcher.max_off_policy_level}"
         )
 
-        # Save final eval rollouts to disk
-        eval_rollouts = [o for outputs in eval_results for o in outputs]
-        if eval_rollouts:
-            step_path = get_step_path(get_rollout_dir(config.output_dir), progress.step)
-            await asyncio.to_thread(
-                save_rollouts, eval_rollouts, step_path / "eval_rollouts.jsonl", exclude_keys={"trajectory"}
+        self.train_sink.reset_pre_filter_stats()
+        self.progress.step += 1
+
+    def log_eval_batch(self, batch: EvalBatch) -> None:
+        """Persist + log one completed eval epoch. Builds the metrics dict
+        from the raw rollouts (the ``EvalBatch`` deliberately doesn't carry a
+        pre-baked metrics dict — it's a pure view computed here), then handles
+        the side effects: save_rollouts, monitor.log_eval_samples, monitor.log."""
+        if not batch.rollouts:
+            self.logger.warning(f"Eval @ step={batch.step} env={batch.env_name}: no surviving rollouts, skipping log")
+            return
+
+        assert self.eval_sink is not None and self.monitor is not None
+        metrics = self.eval_sink.build_metrics(batch)
+
+        step_path = get_step_path(get_rollout_dir(self.config.output_dir), batch.step)
+        save_rollouts(
+            batch.rollouts,
+            step_path / f"eval_rollouts_{batch.env_name}.jsonl",
+            exclude_keys={"trajectory"},
+        )
+        self.monitor.log_eval_samples(batch.rollouts, env_name=batch.env_name, step=batch.step)
+        self.monitor.log(metrics, step=batch.step)
+
+        n_examples = len({r["example_id"] for r in batch.rollouts})
+        reward_mean = metrics.get(f"eval/{batch.env_name}/reward/mean", float("nan"))
+        self.logger.success(
+            f"Eval @ step={batch.step} env={batch.env_name} | "
+            f"Reward: {reward_mean:.4f} | Rollouts: {len(batch.rollouts)} | Examples: {n_examples}"
+        )
+
+    async def maybe_save_ckpt(self, step: int) -> float:
+        """Save the checkpoint if we're at an interval boundary. Returns the
+        elapsed time (0.0 if no save happened)."""
+        if self.ckpt_manager is None or self.config.ckpt is None or not self.config.ckpt.interval:
+            return 0.0
+        if step <= 0:
+            return 0.0
+        is_last_step = self.config.max_steps is not None and step == self.config.max_steps - 1
+        if is_last_step:
+            return 0.0
+        if step % self.config.ckpt.interval != 0:
+            return 0.0
+        self.logger.info(f"Saving checkpoint at step {step}")
+        t = time.perf_counter()
+        await asyncio.to_thread(self.ckpt_manager.save, self.progress, step)
+        return time.perf_counter() - t
+
+    async def wait_barrier(self, step: int) -> None:
+        """Block until ``policy.version >= step - 1`` so the orchestrator stays
+        at most one step ahead of the trainer.
+
+        Cascades backpressure into the dispatcher via the bounded ``out_q``:
+        while we wait, the queue fills, the dispatcher stops handing out new
+        semaphore permits, and in-flight rollouts drain without new ones being
+        scheduled.
+        """
+        target_lag = 1
+        next_warn = 60.0
+        t0 = time.perf_counter()
+        while True:
+            lead = step - self.policy.version
+            if lead <= target_lag:
+                return
+            elapsed = time.perf_counter() - t0
+            if elapsed >= next_warn:
+                # Just a stall observation — no speculation about cause.
+                self.logger.info(
+                    f"Orchestrator waiting at async barrier ({int(elapsed)}s): step={step}, "
+                    f"policy.version={self.policy.version}, lead={lead} (max_async_level={target_lag})."
+                )
+                next_warn = elapsed + 60.0
+            await asyncio.sleep(0.1)
+
+    # ── shutdown ──────────────────────────────────────────────────────────
+
+    async def shutdown(self) -> None:
+        """Bounded best-effort cleanup. Has a global timeout so a wedged peer
+        can't keep the process alive forever — training artifacts are already
+        persisted before this is reached."""
+
+        async def do_shutdown() -> None:
+            if self.sender is not None:
+                self.sender.close()
+            if self.dispatcher is not None:
+                await self.dispatcher.stop()
+            if self.watcher is not None:
+                await self.watcher.stop()
+            if self.log_loop is not None:
+                await self.log_loop.stop()
+            for task in self.component_tasks:
+                await safe_cancel(task)
+            self.component_tasks.clear()
+            if self.inference_metrics is not None:
+                await self.inference_metrics.stop()
+            if self.student_inference is not None:
+                await self.student_inference.stop()
+            if self.teacher_inference is not None:
+                await self.teacher_inference.stop()
+            if self.train_envs is not None:
+                self.train_envs.shutdown()
+            if self.eval_envs is not None:
+                self.eval_envs.shutdown()
+            if self.usage_reporter is not None:
+                self.usage_reporter.close()
+
+        task = asyncio.create_task(do_shutdown())
+        _, pending = await asyncio.wait({task}, timeout=SHUTDOWN_TIMEOUT_S)
+        if pending:
+            self.logger.warning(
+                f"Orchestrator shutdown did not complete within {SHUTDOWN_TIMEOUT_S}s; "
+                "forcing process exit. Training artifacts are already persisted."
             )
-
-    monitor.save_final_summary()
-
-    # Write final checkpoint
-    if ckpt_manager is not None:
-        logger.info("Writing final checkpoint")
-        ckpt_manager.save(progress, buffer, step=progress.step)
-
-    # Bounded best-effort cleanup. Each await below may block on a remote peer
-    # (env-server ZMQ recv, inference admin httpx aclose, etc.). The outer
-    # asyncio.wait gives the whole sequence a single deadline; if anything
-    # wedges past SHUTDOWN_TIMEOUT_S we force-exit the process. Individual
-    # awaits intentionally do NOT have their own timeouts — asyncio.wait_for
-    # would itself hang on an uncancellable await, which is exactly the
-    # failure mode we're guarding against.
-    async def _graceful_shutdown() -> None:
-        training_batch_sender.close()
-        await scheduler.stop()
-        if inference_metrics_collector is not None:
-            await inference_metrics_collector.stop()
-        await student_inference.stop()
-        if teacher_inference is not None:
-            await teacher_inference.stop()
-        event_loop_lag_monitor_task.cancel()
-        # Shutdown env processes (also registered as atexit handler for crash safety)
-        train_envs.shutdown()
-        if eval_envs is not None:
-            eval_envs.shutdown()
-
-    shutdown_task = asyncio.create_task(_graceful_shutdown())
-    _, pending = await asyncio.wait({shutdown_task}, timeout=SHUTDOWN_TIMEOUT_S)
-
-    if pending:
-        logger.warning(
-            f"Orchestrator shutdown did not complete within {SHUTDOWN_TIMEOUT_S}s; "
-            "forcing process exit. Training artifacts are already persisted."
-        )
-        os._exit(0)
-
-    # asyncio.wait swallows task exceptions; re-raise so a fast cleanup
-    # failure surfaces the same way as it did when each step was awaited
-    # directly.
-    await shutdown_task
-
-    if usage_reporter:
-        usage_reporter.close()
-
-    logger.success("Orchestrator finished.")
-
-    # Optionally, print benchmark table
-    if config.bench:
-        print_benchmark(to_col_format(monitor.history))
+            os._exit(0)
+        await task
 
 
-def main():
-    """Main entry-point for orchestrator. Run using `uv run orchestrator`"""
-    set_proc_title("Orchestrator")
-    import uvloop
+@clean_exit
+async def run_orchestrator(config: OrchestratorConfig) -> None:
+    """Top-level entrypoint: instantiate ``Orchestrator`` and run it.
 
-    uvloop.install()
-    asyncio.run(orchestrate(cli(OrchestratorConfig)))
-
-
-async def setup_student_inference_pool(
-    *,
-    config: OrchestratorConfig,
-    tokenizer,
-    logger,
-):
-    """Set up the student inference pool (rollouts when rl/opd, evals + weight sync always).
-
-    Routing policy is driven by ``config.use_renderer``:
-
-      - ``use_renderer=True``  → renderer-backed TITO client (``/v1/generate``).
-        Default for both text-only and VLM rollouts; required for VLMs.
-      - ``use_renderer=False`` → MITO (``openai_chat_completions``).
-
-    Eval clients always use MITO. In sft mode ``use_renderer`` is forced off
-    by a config validator, so the student pool is plain MITO end-to-end.
+    Wrapped in ``@clean_exit`` so wandb is flushed on exit (success or crash);
+    keeps that cleanup out of the ``Orchestrator`` class proper.
     """
+    await Orchestrator(config).start()
+
+
+async def setup_student_inference_pool(*, config: OrchestratorConfig, tokenizer):
+    """Mirror of ``prime_rl.orchestrator.orchestrator.setup_student_inference_pool``."""
+    from renderers.base import create_renderer
+
     client_config = config.student.client
     model_name = config.student.model.name
 
@@ -880,7 +667,7 @@ async def setup_student_inference_pool(
             preserve_all_thinking=config.renderer.preserve_all_thinking,
             preserve_thinking_between_tool_calls=config.renderer.preserve_thinking_between_tool_calls,
         )
-        logger.info(f"Initialized {type(renderer).__name__} for {model_name}")
+        get_logger().info(f"Initialized {type(renderer).__name__} for {model_name}")
         inference_pool = await setup_inference_pool(
             client_config,
             model_name=model_name,
@@ -893,10 +680,10 @@ async def setup_student_inference_pool(
             preserve_all_thinking=config.renderer.preserve_all_thinking,
             preserve_thinking_between_tool_calls=config.renderer.preserve_thinking_between_tool_calls,
         )
-        logger.info("Using direct renderer rollout client")
+        get_logger().info("Using direct renderer rollout client")
         return renderer, inference_pool
 
-    logger.info("Using MITO (openai_chat_completions) for rollouts")
+    get_logger().info("Using MITO (openai_chat_completions) for rollouts")
     inference_pool = await setup_inference_pool(
         client_config,
         model_name=model_name,
@@ -904,6 +691,18 @@ async def setup_student_inference_pool(
         eval_client_type="openai_chat_completions",
     )
     return None, inference_pool
+
+
+def main() -> None:
+    """Entrypoint for direct invocation via ``python -m prime_rl.orchestrator.orchestrator``."""
+    from prime_rl.utils.config import cli
+    from prime_rl.utils.process import set_proc_title
+
+    set_proc_title("OrchestratorV2")
+    import uvloop
+
+    uvloop.install()
+    asyncio.run(run_orchestrator(cli(OrchestratorConfig)))
 
 
 if __name__ == "__main__":
